FPS Improvements - CFG_Init has a major flaw, do not use with INI_FIND_VALUE

+++ ini_parser.c
@@ -20,11 +20,11 @@ static inline unsigned TO_UINT(const char* str) {
 static inline unsigned int INI_FIND_VALUE(const char *key) {
     for (int i = 0; i < INI_ENTRY_COUNT; ++i) {
         if (strcmp(INIENTRIES[i].key, key) == 0) {
-            return INIENTRIES[i].value;
+            return (int)INIENTRIES[i].value;
         }
     }
     fprintf(stderr, "Failed to find key \"%s\"\n", key);
-    return 0;  // or some sentinel
+    return 0;
 }
 
 void INI_PARSE(const char* filename) { // File must be within the main folder and not hidden deep within a subfolder.
@@ -40,11 +40,7 @@ void INI_PARSE(const char* filename) { // File must be within the main folder an
     while (fgets(line, sizeof(line), f)) {
         char *p = line;
         while (*p == ' ' || *p == '\t') ++p;
-
-        // Skip comments and empty lines
         if (*p == '#' || *p == ';' || *p == '\n') continue;
-
-        // Section header
         if (*p == '[') {
             char* end = strchr(p, ']');
             if (end) {
@@ -53,17 +49,13 @@ void INI_PARSE(const char* filename) { // File must be within the main folder an
             }
             continue;
         }
-
         char* equal = strchr(p, '=');
         if (equal) {
             *equal = 0;
             char* key = p;
             char* value = equal + 1;
-
-            // Trim newline from value
             char* newline = strchr(value, '\n');
             if (newline) *newline = 0;
-
             if (INI_ENTRY_COUNT < MAX_ENTRIES) {
                 INIENTRIES[INI_ENTRY_COUNT++] = (iniEntry){
                     .section = strdup(current_section),
